New sizes for the new ATLAS logo. Details: https://its.cern.ch/jira/browse/ATLASVPONE-214 (VP1Gui-00-08-03)
Former-commit-id: afccaa9589cf042164d389e89f392b591a5a1dbc

+++ graphics/VP1/VP1Gui/VP1Gui/VP1EventDisplayEditor.h
@@ -37,7 +37,7 @@ class VP1EventDisplayEditor : public QWidget {
 
 public:
 
-	VP1EventDisplayEditor(VP1MainWindow* mainWindow, QList<unsigned long long> listRunEventNumberTimestamp);
+	VP1EventDisplayEditor(VP1MainWindow* mainWindow, QList<int> listRunEventNumberTimestamp);
 	virtual ~VP1EventDisplayEditor();
 
 //	QGraphicsView* getView();
@@ -83,7 +83,7 @@ private:
 	VP1MainWindow* _mainWindow;
 
 	int _runNumber;
-	unsigned long long _eventNumber;
+	int _eventNumber;
 	unsigned _eventTimestamp;
 
 //	QSplitter *h1Splitter;
