+++ graphics/VP1/VP1Gui/VP1Gui/VP1Gui.h
@@ -19,7 +19,6 @@
 
 #include <string>
 #include <vector>
-#include <cstdint> // for uint64_t (unsigned long long)
 
 class StoreGateSvc;
 class IToolSvc;
@@ -46,7 +45,7 @@ public:
 
   //Called every algorithm refresh:
   bool executeNewEvent( const int& run,
-			const uint64_t & event,
+			const int& event,
 			const unsigned& triggerType = 0,//When available
 			const unsigned& time = 0 //When available (seconds since 1970)
 			);
